requirements.txt
tensorflow-gpu==1.15.4
keras==2.2.4
numpy==1.16
pandas==0.23.4
seaborn==0.9.0
matplotlib==3.3.2
scikit-image==0.16.2
SimpleITK==2.0
opencv-python==4.1.2.30
nibabel==3.0.0
h5py==2.10